support all types of capacitors
