+++ src/main/java/com/enderio/base/common/item/capacitors/LootCapacitorItem.java
@@ -5,12 +5,11 @@ import com.enderio.api.capability.MultiCapabilityProvider;
 import com.enderio.base.common.capacitor.LootCapacitorData;
 import com.enderio.base.common.init.EIOCapabilities;
 import net.minecraft.nbt.CompoundTag;
-import net.minecraft.world.item.Item;
 import net.minecraft.world.item.ItemStack;
 import net.minecraftforge.common.util.LazyOptional;
 import org.jetbrains.annotations.Nullable;
 
-public class LootCapacitorItem extends Item implements IMultiCapabilityItem {
+public class LootCapacitorItem extends BaseCapacitorItem implements IMultiCapabilityItem {
 
     public LootCapacitorItem(Properties pProperties) {
         super(pProperties);
